pyproject.toml
[project]
name = "project2-studypal-study-assistance"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "httpx>=0.28.1",
    "langchain-chroma>=1.1.0",
    "langchain-community>=0.4.1",
    "langchain-groq>=1.1.2",
    "langchain-huggingface>=1.2.0",
    "langchain-text-splitters>=1.1.0",
    "nltk>=3.9.2",
    "pdf2image>=1.17.0",
    "pillow>=12.1.1",
    "pytesseract>=0.3.13",
    "python-dotenv>=1.2.1",
    "sentence-transformers>=5.2.2",
    "streamlit>=1.54.0",
    "transformers>=4.57.6",
    "unstructured[pdf]>=0.20.2",
    "youtube-search-python>=1.6.6",
]
